New improved service `heatmap` substitute of `mapPlace` and `locations`
